refactor(core): rename par module to parallel; simplify CLI bin cfg
- Rename src/par.rs to src/parallel.rs (clearer name for the rayon/sequential
  abstraction) and update its references.
- Collapse the per-line wasm cfg guards in main.rs to a single unconditional
  main() that calls a cfg-gated run() holding the native-only imports.
