+++ src/parallel.rs
